+++ examples/esp8266_example.cpp
@@ -3,10 +3,20 @@
  */
 
 #include <Arduino.h>
-#include "NeoPixelEspEx.h"
 #include <core_version.h>
+#include "NeoPixelEspEx.h"
 
-NeoPixelEx::Strip<NEOPIXEL_OUTPUT_PIN, NEOPIXEL_NUM_PIXELS, NeoPixelEx::GRB, NEOPIXEL_CHIPSET> pixels;
+// #define NEOPIXEL_NUM_PIXELS_CLEAR NEOPIXEL_NUM_PIXELS * 2
+#define NEOPIXEL_NUM_PIXELS_CLEAR NEOPIXEL_NUM_PIXELS + 10
+
+using namespace NeoPixelEx;
+
+Strip<NEOPIXEL_OUTPUT_PIN, NEOPIXEL_NUM_PIXELS, GRB, NEOPIXEL_CHIPSET> pixels;
+
+// example of using the DataWrapper to use any raw pointer with the Strip class
+//
+// uint8_t pixelData[NEOPIXEL_NUM_PIXELS * sizeof(GRB)];
+// Strip<NEOPIXEL_OUTPUT_PIN, NEOPIXEL_NUM_PIXELS, GRB, NEOPIXEL_CHIPSET, DataWrapper<NEOPIXEL_NUM_PIXELS, GRB>> pixels(&pixelData);
 
 extern bool record_serial_output;
 extern String serial_output;
@@ -50,8 +60,7 @@ void setup()
     ::printf("setup\n");
     uart_uninit(_uart);
 #endif
-    NeoPixelEx::forceClear<decltype(pixels)::chipset_type>(pixels.kOutputPin, 256);
-    delay(2);
+    forceClear<decltype(pixels)::chipset_type>(pixels.kOutputPin, NEOPIXEL_NUM_PIXELS_CLEAR);
 
     Serial.begin(115200);
     Serial.println(F("Starting..."));
@@ -60,9 +69,9 @@ void setup()
     pixels.begin();
 }
 
-NeoPixelEx::GRB color(0, 0, 0xff);
+GRB color(0, 0, 0xff);
 uint8_t brightness = 10;
-uint8_t delayMillis = 1;
+uint8_t delayMillis = 4;
 uint8_t *valuePtr = &brightness;
 
 static PGM_P get_type_str()
@@ -114,7 +123,7 @@ static int delayMillisOld = -1;
 
 void end_test(const __FlashStringHelper *msg)
 {
-    NeoPixelEx::forceClear<decltype(pixels)::chipset_type>(pixels.kOutputPin, 256);
+    forceClear<decltype(pixels)::chipset_type>(pixels.kOutputPin, NEOPIXEL_NUM_PIXELS_CLEAR);
     testAll = -1;
     if (delayMillisOld != -1) {
         delayMillis = delayMillisOld;
@@ -129,7 +138,7 @@ void loop()
 
     if (testAll != -1 && counter > 100) {
         counter = 0;
-        using ColorType = NeoPixelEx::GRB;
+        using ColorType = GRB;
         // test pixels
         static ColorType onColor;
 
@@ -161,19 +170,29 @@ void loop()
         ColorType offColor = onColor.inverted();
         offColor.setBrightness(128);
         #if NEOPIXEL_HAVE_STATS
-            auto &stats = NeoPixelEx::getStats();
-            Serial.printf_P(PSTR("testing color on=%s off=%s pixel=%u/%u run=%u dropped=%u frames=%u fps=%.2f delay=10\n"), onColor.toString().c_str(), offColor.toString().c_str(), currentPixelNum, pixels.getNumPixels(), numRun, stats.getAbortedFrames(), stats.getFrames(), stats.getFps(), delayMillis);
+            auto &stats = pixels.getStats();
+            Serial.printf_P(PSTR("testing color on=%s off=%s pixel=%u/%u run=%u dropped=%u frames=%u fps=%u delay=10\n"), onColor.toString().c_str(), offColor.toString().c_str(), currentPixelNum, pixels.getNumPixels(), numRun, stats.getAbortedFrames(), stats.getFrames(), stats.getFps(), delayMillis);
         #else
 
         #endif
         pixels.fill(offColor);
         pixels[currentPixelNum] = onColor;
-        pixels.show(20);
+        pixels.show(10);
         testAll++;
     }
     else if (testAll == -1) {
         // display pixels
         pixels.fill(color);
+
+        // GRB pattern[5] = { color, color.scale(200), color.scale(100), color.scale(75), color.scale(25) };
+        // auto begin = static_cast<uint8_t *>(pixels);
+        // auto end = begin + pixels.getNumBytes();
+        // while(begin + sizeof(pattern) < end) {
+        //     memcpy(begin, pattern, sizeof(pattern));
+        //     begin += sizeof(pattern);
+        // }
+        // std::fill(begin, end, 0);
+
         pixels.show(brightness);
     }
 
@@ -185,6 +204,31 @@ void loop()
         while(Serial.available()) {
             char ch  = Serial.read();
             switch(ch) {
+                case 'M': {
+                        pixels.fill(0);
+                        auto &data = pixels.data();
+                        auto endPtr = data.end();
+                        for(auto ptr = data.begin(); ptr < endPtr; ptr += 10) {
+                            *ptr = 0x000011;
+                        }
+                        for(auto ptr = data.begin() + 4; ptr < endPtr; ptr += 10) {
+                            *ptr = 0x000044;
+                        }
+                        for(auto ptr = data.begin() + 9; ptr < endPtr; ptr += 10) {
+                            *ptr = 0x0f0088;
+                        }
+                        for(auto ptr = data.begin() + 49; ptr < endPtr; ptr += 50) {
+                            *ptr = 0x004400;
+                        }
+                        for(auto ptr = data.begin() + 99; ptr < endPtr; ptr += 100) {
+                            *ptr = 0x008800;
+                        }
+                        pixels[pixels.size() - 1] = 0xff0000;
+                        pixels.show(255);
+                        delay(30000);
+                        pixels.clear();
+                    }
+                    break;
                 case 't':
                     if (delayMillisOld == -1) {
                         delayMillisOld = delayMillis;
@@ -195,7 +239,7 @@ void loop()
                     testAll = 0;
                     counter = -1;
                     #if NEOPIXEL_HAVE_STATS
-                        NeoPixelEx::getStats().clear();
+                        pixels.getStats().clear();
                     #endif
                     pixels.clear();
                     break;
@@ -226,31 +270,57 @@ void loop()
                 case 'd':
                     valuePtr = &delayMillis;
                     break;
+                case '1':
+                    if (pixels[0] != 0) {
+                        color = 0;
+                    }
+                    else {
+                        color = 0xffffff;
+                    }
+                    break;
+                case '2':
+                    if (pixels[0] != 0) {
+                        brightness = 0;
+                    }
+                    else {
+                        brightness = 255;
+                    }
+                    break;
+                case '3':
+                    if (pixels[0] != 0) {
+                        brightness = 0;
+                        color = 0;
+                    }
+                    else {
+                        brightness = 255;
+                        color = 0xffffff;
+                    }
+                    break;
                 case 'S':
-                    Serial.println(F("clear stat..."));
+                    Serial.println(F("clearing stats..."));
                     #if NEOPIXEL_HAVE_STATS
-                        NeoPixelEx::getStats().clear();
+                        pixels.getStats().clear();
                     #endif
                     break;
                 case 'R':
                     Serial.println(F("reset"));
-                    color = NeoPixelEx::GRB(0, 0, 0xff);
+                    color = GRB(0, 0, 0xff);
                     brightness = 16;
                     delayMillis = 1;
                     valuePtr = &brightness;
                     // falltrough
                 case 'C':
                     Serial.println(F("clearing all pixels..."));
-                    NeoPixelEx::forceClear<decltype(pixels)::chipset_type>(pixels.kOutputPin, 256);
+                    forceClear<decltype(pixels)::chipset_type>(pixels.kOutputPin, NEOPIXEL_NUM_PIXELS_CLEAR);
                     Serial.println(F("waiting 3 seconds..."));
                     delay(3000);
                     clear_keys();
                     #if NEOPIXEL_HAVE_STATS
-                        NeoPixelEx::getStats().clear();
+                        pixels.getStats().clear();
                     #endif
                     break;
                 case 'H':
-                    NeoPixelEx::forceClear<decltype(pixels)::chipset_type>(pixels.kOutputPin, 256);
+                    forceClear<decltype(pixels)::chipset_type>(pixels.kOutputPin, NEOPIXEL_NUM_PIXELS_CLEAR);
                     Serial.println(F("halted..."));
                     for(;;) {
                         delay(1000);
@@ -263,8 +333,8 @@ void loop()
                     break;
             }
             #if NEOPIXEL_HAVE_STATS
-                const auto &stats = NeoPixelEx::getStats();
-                Serial.printf_P(PSTR("selected=%s step=%u brightness=%u %s dropped=%u frames=%u fps=%.2f delay=%u\n"), get_type_str(), get_type_incr(), brightness, color.toString().c_str(), stats.getAbortedFrames(), stats.getFrames(), stats.getFps(), delayMillis);
+                const auto &stats = pixels.getStats();
+                Serial.printf_P(PSTR("selected=%s step=%u brightness=%u %s dropped=%u frames=%u fps=%u delay=%u\n"), get_type_str(), get_type_incr(), brightness, color.toString().c_str(), stats.getAbortedFrames(), stats.getFrames(), stats.getFps(), delayMillis);
             #else
                 Serial.printf_P(PSTR("selected=%s step=%u brightness=%u %s delay=%u\n"), get_type_str(), get_type_incr(), brightness, color.toString().c_str(), delayMillis);
             #endif