+++ include/NeoPixelEspEx.h
@@ -12,7 +12,7 @@
 
 // enable debug mode
 #ifndef NEOPIXEL_DEBUG
-#define NEOPIXEL_DEBUG 1
+#define NEOPIXEL_DEBUG 0
 #endif
 
 // enable the brightness scaling
@@ -76,11 +76,14 @@
 #   define NEOPIXEL_DEBUG_TRIGGER_PIN -1
 #endif
 
+#pragma GCC push_options
+#pragma GCC optimize ("O2")
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-bool espShow(uint8_t pin, uint16_t brightness, const uint8_t *p, const uint8_t *end, uint32_t time0, uint32_t time1, uint32_t period, uint32_t wait, uint32_t minWaitPeriod, uint32_t pinMask, uint32_t *lastDisplayTime);
+bool espShow(uint8_t pin, uint16_t brightness, const uint8_t *p, const uint8_t *end, uint32_t time0, uint32_t time1, uint32_t period, uint32_t wait, uint32_t minWaitPeriod, uint32_t pinMask, void *context);
 
 // deprecated legacy function
 // fill pixel data
@@ -91,44 +94,13 @@ void NeoPixel_fillColor(uint8_t *pixels, uint16_t numBytes, uint32_t color);
 // show pixels
 // numBytes is 3 per pixel
 // brightness 0-255
-bool NeoPixel_espShow(uint8_t pin, const uint8_t *pixels, uint16_t numBytes, uint16_t brightness = 255);
+// context is a pointer to thje context object of the LED strip. nullptr will use the shared global object
+bool NeoPixel_espShow(uint8_t pin, const uint8_t *pixels, uint16_t numBytes, uint16_t brightness = 255, void *context = nullptr);
 
 #ifdef __cplusplus
 }
 
 namespace NeoPixelEx {
-
-    // no retries bit for brightness
-    #if NEOPIXEL_DEBUG_TRIGGER_PIN==16
-        #error GPIO16 not supported
-    #elif NEOPIXEL_DEBUG_TRIGGER_PIN>=0
-    extern bool _toogleDebugFlag;
-
-    __attribute__((always_inline)) inline bool NeoPixel_initToggleDebugPin()
-    {
-        digitalWrite(NEOPIXEL_DEBUG_TRIGGER_PIN, LOW);
-        pinMode(NEOPIXEL_DEBUG_TRIGGER_PIN, OUTPUT);
-        NeoPixelEx::_toogleDebugFlag = false;
-        return false;
-    }
-
-    __attribute__((always_inline)) inline void NeoPixel_toggleDebugPin()
-    {
-        if ((_toogleDebugFlag = !_toogleDebugFlag)) {
-            GPOC = _BV(NEOPIXEL_DEBUG_TRIGGER_PIN);
-        } else {
-            GPOS = _BV(NEOPIXEL_DEBUG_TRIGGER_PIN);
-        }
-    }
-    #else
-    __attribute__((always_inline)) inline bool NeoPixel_initToggleDebugPin() {
-        return false;
-    }
-
-    __attribute__((always_inline)) inline void NeoPixel_toggleDebugPin() {
-    }
-    #endif
-
 
     // timits in nano seconds
     template<uint32_t _T0H, uint32_t _T1H, uint32_t _TPeriod, uint32_t _TReset, uint32_t _FCpu/*Hz or MHz*/>
@@ -195,10 +167,6 @@ namespace NeoPixelEx {
 
     class Stats {
     public:
-        Stats() : _start(millis64()), _frames(0)
-        {
-        }
-
         void clear() {
             *this = Stats();
         }
@@ -207,8 +175,9 @@ namespace NeoPixelEx {
             return _frames;
         }
 
-        float getFps() const {
-            return (_frames * 1000UL) / static_cast<float>(getTime());
+        uint16_t getFps() const {
+            auto time = getTime();
+            return time ? (_frames * 1000UL) / getTime() : 0;
         }
 
         // time since last reset in milliseconds
@@ -233,7 +202,11 @@ namespace NeoPixelEx {
     public:
     #if NEOPIXEL_ALLOW_INTERRUPTS
 
-        __attribute__((always_inline)) uint32_t getAbortedFrames() const {
+        __attribute__((always_inline)) inline Stats() : _start(millis64()), _frames(0), _aborted(0)
+        {
+        }
+
+        __attribute__((always_inline)) inline uint32_t getAbortedFrames() const {
             return _aborted;
         }
 
@@ -241,14 +214,26 @@ namespace NeoPixelEx {
             return true;
         }
 
-    public:
-        __attribute__((always_inline)) uint32_t &__aborted() {
-            return _aborted;
+        __attribute__((always_inline)) inline void increment(bool success) {
+            _frames++;
+            if (success) {
+                return;
+            }
+            _aborted++;
         }
 
     private:
-        uint32_t _aborted{0};
+        uint32_t _aborted;
     #else
+
+    public:
+        __attribute__((always_inline)) inline Stats() : _start(millis64()), _frames(0)
+        {
+        }
+
+        __attribute__((always_inline)) inline void increment(bool success) {
+            _frames++;
+        }
 
         static constexpr uint32_t getAbortedFrames() {
             return 0;
@@ -259,6 +244,115 @@ namespace NeoPixelEx {
         }
 
     #endif
+    };
+
+    class DebugContext {
+    public:
+    static constexpr uint8_t kInvalidPin = 0xff;
+
+    public:
+        DebugContext(uint8_t pin = kInvalidPin) :
+            _pin(kInvalidPin),
+            _enabled(false),
+            _state(false)
+        {
+            begin(pin);
+        }
+
+        void begin(uint8_t pin) {
+            end();
+            _pin = pin;
+            if (_pin != kInvalidPin) {
+                digitalWrite(_pin, _state);
+                pinMode(_pin, OUTPUT);
+                _enabled = true;
+            }
+        }
+
+        void end() {
+            _state = false;
+            _enabled = false;
+            if (_pin != kInvalidPin) {
+                digitalWrite(_pin, _state);
+                pinMode(_pin, INPUT);
+                _pin = kInvalidPin;
+            }
+        }
+
+        void togglePin() {
+            if (_enabled) {
+                _state = !_state;
+                #if defined(ESP8266)
+                    if (_state) {
+                        GPOC = _BV(_pin);
+                    }
+                    else {
+                        GPOS = _BV(_pin);
+                    }
+                #else
+                    digitalWrite(_pin, state);
+                #endif
+            }
+        }
+
+    private:
+        uint8_t _pin;
+        bool _enabled;
+        bool _state;
+    };
+
+    class Context {
+    public:
+        Context() :
+            #if NEOPIXEL_DEBUG && (NEOPIXEL_DEBUG_TRIGGER_PIN >= 0)
+                _debug(NEOPIXEL_DEBUG_TRIGGER_PIN),
+            #endif
+            _lastDisplayTime(0)
+        {
+        }
+
+        uint32_t &getLastDisplayTime() {
+            return _lastDisplayTime;
+        }
+
+        uint32_t getLastDisplayTime() const {
+            return _lastDisplayTime;
+        }
+
+        void setLastDisplayTime(uint32_t micros) {
+            _lastDisplayTime = micros;
+        }
+
+        void waitRefreshTime(uint32_t minWaitPeriod) {
+            uint32_t diff = micros() - getLastDisplayTime();
+            if (diff < minWaitPeriod) {
+                delayMicroseconds(minWaitPeriod - diff);
+            }
+        }
+
+    #if NEOPIXEL_HAVE_STATS
+        Stats &getStats() {
+            return _stats;
+        }
+    #endif
+
+    #if NEOPIXEL_DEBUG
+        DebugContext &getDebugContext() {
+            return _debug;
+        }
+    #endif
+
+        // returns global context if contextPtr is nullptr
+        static Context &validate(void *contextPtr);
+
+    private:
+    #if NEOPIXEL_HAVE_STATS
+        Stats _stats;
+    #endif
+    #if NEOPIXEL_DEBUG
+        DebugContext _debug;
+    #endif
+        uint32 _lastDisplayTime;
     };
 
     class GRBType
@@ -390,6 +484,14 @@ namespace NeoPixelEx {
             }
         }
 
+        Color scale(uint8_t brightness) const {
+            if (brightness == 0) {
+                return Color(0);
+            }
+            brightness++;
+            return Color((red() * brightness) >> 8, (green() * brightness) >> 8, (blue() * brightness) >> 8);
+        }
+
         String toString() const {
             char buf[10];
             snprintf_P(buf, sizeof(buf), PSTR("#%06x"), toRGB());
@@ -414,30 +516,106 @@ namespace NeoPixelEx {
     using GRB = Color<GRBType>;
     using RGB = Color<RGBType>;
 
-    template<uint16_t _NumPixels, typename _PixelType = GRB>
-    class Array : public std::array<_PixelType, _NumPixels>
+    // wrapper for any raw pointer
+    template<size_t _NumElements, typename _PixelType = GRB>
+    class DataWrapper {
+    public:
+        DataWrapper(void *data) : _data(reinterpret_cast<_PixelType *>(data)) {}
+
+        static constexpr size_t size() {
+            return _NumElements;
+        }
+
+        void fill(_PixelType color) {
+            std::fill_n(reinterpret_cast<_PixelType *>(_data), size(), color);
+        }
+
+        _PixelType *data() {
+            return _data;
+        }
+
+        const _PixelType *data() const {
+            return _data;
+        }
+
+        _PixelType *begin() {
+            return _data;
+        }
+
+        const _PixelType *begin() const {
+            return _data;
+        }
+
+        _PixelType *end() {
+            return &_data[size()];
+        }
+
+        const _PixelType *end() const {
+            return &_data[size()];
+        }
+
+        _PixelType &operator[](int index) {
+            return _data[index];
+        }
+
+        _PixelType operator[](int index) const {
+            return _data[index];
+        }
+
+    private:
+        _PixelType *_data;
+    };
+
+    template<uint16_t _NumPixels, typename _PixelType = GRB, typename _DataType = std::array<_PixelType, _NumPixels>>
+    class PixelData
     {
     public:
-        using type = std::array<_PixelType, _NumPixels>;
+        using data_type = std::array<_PixelType, _NumPixels>;
         using pixel_type = _PixelType;
-        using type::data;
-        using type::size;
-        using type::begin;
-        using type::end;
 
         static constexpr uint16_t kNumPixels = _NumPixels;
         static constexpr uint16_t kNumBytes = kNumPixels * sizeof(pixel_type);
 
-        static constexpr uint16_t kGetNumBytes(const uint16_t offset = 0, const uint16_t num = kNumPixels) {
-            return (num - offset) * sizeof(pixel_type);
-        }
+    public:
+        PixelData() {}
 
-        constexpr uint16_t getNumBytes() const {
+        template<typename ..._Args>
+        PixelData(_Args &&...args) : _data(std::forward<_Args &&>(args)...) {}
+
+        static  constexpr uint16_t getNumBytes() {
             return kNumBytes;
         }
 
-        constexpr uint16_t getNumPixels() const {
+        static constexpr uint16_t getNumPixels() {
             return kNumPixels;
+        }
+
+        static constexpr uint16_t size() {
+            return kNumPixels;
+        }
+
+        pixel_type *data() {
+            return _data.data();
+        }
+
+        const pixel_type *data() const {
+            return _data.data();
+        }
+
+        pixel_type *begin() {
+            return data();
+        }
+
+        const pixel_type *begin() const {
+            return data();
+        }
+
+        pixel_type *end() {
+            return &data()[_data.size()];
+        }
+
+        const pixel_type *end() const {
+            return data()[_data.size()];
         }
 
         explicit operator const pixel_type *() const {
@@ -456,54 +634,54 @@ namespace NeoPixelEx {
             return reinterpret_cast<uint8_t *>(data());
         }
 
-        inline void set(int index, pixel_type color) {
+        void set(int index, pixel_type color) {
             data()[index] = color;
         }
 
-        inline pixel_type get(int index) const {
+        pixel_type get(int index) const {
             return data()[index];
         }
 
-        inline pixel_type &operator[](int index) {
+        pixel_type &operator[](int index) {
             return data()[index];
         }
 
-        inline pixel_type operator[](int index) const {
+        pixel_type operator[](int index) const {
             return data()[index];
         }
 
-        inline void fill(uint32_t color) {
+        void fill(uint32_t color) {
             std::fill(begin(), end(), pixel_type(color));
         }
 
-        inline void fill(const pixel_type &color) {
+        void fill(const pixel_type &color) {
             std::fill(begin(), end(), color);
         }
 
-        inline void fill(uint16_t numPixels, uint32_t color) {
+        void fill(uint16_t numPixels, uint32_t color) {
             std::fill_n(data(), numPixels, pixel_type(color));
         }
 
-        inline void fill(uint16_t numPixels, const pixel_type &color) {
+        void fill(uint16_t numPixels, const pixel_type &color) {
             std::fill_n(data(), numPixels, color);
         }
 
-        inline void fill(uint16_t offset, uint16_t numPixels, uint32_t color) {
+        void fill(uint16_t offset, uint16_t numPixels, uint32_t color) {
             std::fill(data() + offset, data() + numPixels, pixel_type(color));
         }
 
-        inline void fill(uint16_t offset, uint16_t numPixels, const pixel_type &color) {
+        void fill(uint16_t offset, uint16_t numPixels, const pixel_type &color) {
             std::fill(data() + offset, data() + numPixels, color);
         }
+
+    private:
+        data_type _data;
     };
 
 
     template<typename _Chipset = NEOPIXEL_CHIPSET>
-    inline bool internalShow(uint8_t pin, const uint8_t *pixels, uint16_t numBytes, uint8_t brightness, uint32_t *lastDisplayTime)
+    inline bool internalShow(uint8_t pin, const uint8_t *pixels, uint16_t numBytes, uint8_t brightness, Context &context)
     {
-    #if 0
-        NeoPixel_toggleDebugPin();
-    #endif
         bool result = false;
         auto p = pixels;
         auto end = p + numBytes;
@@ -515,33 +693,14 @@ namespace NeoPixelEx {
                 #if defined(ESP8266) && !NEOPIXEL_ALLOW_INTERRUPTS
                     ets_intr_lock();
                 #endif
-                result = ::espShow(pin,
-                    brightness,
-                    p,
-                    end,
-                    _Chipset::getCyclesT0H(),
-                    _Chipset::getCyclesT1H(),
-                    _Chipset::getCyclesPeriod(),
-                    _Chipset::getCyclesRES(),
-                    _Chipset::getMinDisplayPeriod(),
-                    pin == 16 ? _BV(0) : _BV(pin),
-                    lastDisplayTime
-                );
+                result = ::espShow(pin, brightness, p, end, _Chipset::getCyclesT0H(), _Chipset::getCyclesT1H(), _Chipset::getCyclesPeriod(), _Chipset::getCyclesRES(), _Chipset::getMinDisplayPeriod(), pin == 16 ? _BV(0) : _BV(pin), &context);
 
                 #if defined(ESP8266) && !NEOPIXEL_ALLOW_INTERRUPTS
                     ets_intr_unlock();
                 #endif
         #if NEOPIXEL_INTERRUPT_RETRY_COUNT > 0
-                if (result == true || lastDisplayTime == nullptr || retries == 0) {
-                    break;
-                }
-                retries--;
-                uint32_t diff = micros() - *lastDisplayTime;
-                if (diff < _Chipset::kMinDisplayPeriod) {
-                    delayMicroseconds(_Chipset::kMinDisplayPeriod - diff);
-                }
             }
-            while(true);
+            while(result != true && retries-- > 0);
         #endif
 
         return result;
@@ -549,18 +708,18 @@ namespace NeoPixelEx {
 
     // force to clear all pixels without interruptions
     template<typename _Chipset = NEOPIXEL_CHIPSET>
-    inline void forceClear(uint8_t pin, uint16_t numBytes)
+    inline void forceClear(uint8_t pin, uint16_t numPixels, Context *contextPtr = nullptr)
     {
         digitalWrite(pin, LOW);
         pinMode(pin, OUTPUT);
-        uint8_t buf[1];
-        uint32_t start = micros();
-        while(micros() - start < DefaultTimings::kMinDisplayPeriod) {
-        }
+
         #if defined(ESP8266) && NEOPIXEL_ALLOW_INTERRUPTS
             ets_intr_lock();
         #endif
-        internalShow<_Chipset>(pin, buf, numBytes, 0, nullptr);
+
+        uint8_t buf[1];
+        internalShow<_Chipset>(pin, buf, numPixels * sizeof(GRB), 0, Context::validate(contextPtr));
+
         #if defined(ESP8266) && NEOPIXEL_ALLOW_INTERRUPTS
             ets_intr_unlock();
         #endif
@@ -570,24 +729,28 @@ namespace NeoPixelEx {
         NeoPixelEx::Stats &getStats();
     #endif
 
-    template<uint8_t _OutputPin, uint16_t _NumPixels, typename _PixelType = GRB, typename _Chipset = TimingsWS2812<F_CPU>>
-    class Strip : public Array<_NumPixels, _PixelType>
+    // _DataType must provide enough data for _NumPixels * sizeof(_PixelType)
+    template<uint8_t _OutputPin, uint16_t _NumPixels, typename _PixelType = GRB, typename _Chipset = TimingsWS2812<F_CPU>, typename _DataType = PixelData<_NumPixels, _PixelType>>
+    class Strip : public PixelData<_NumPixels, _PixelType>
     {
     public:
         static constexpr auto kOutputPin = _OutputPin;
         using chipset_type = _Chipset;
+        using data_type = _DataType;
+        using pixel_type = _PixelType;
 
     public:
-        using type = Array<_NumPixels, _PixelType>;
-        using type::data;
-        using type::fill;
-        using type::getNumBytes;
-        using type::getNumPixels;
-        using type::get;
-        using type::set;
-        using type::operator[];
+        Strip() {}
 
-        Strip() : _lastDisplayTime(micros()) {
+        template<typename ..._Args>
+        Strip(_Args &&...args) : _data(std::forward<_Args &&>(args)...) {}
+
+        constexpr uint16_t getNumBytes() const {
+            return _NumPixels * sizeof(pixel_type);
+        }
+
+        constexpr uint16_t getNumPixels() const {
+            return _NumPixels;
         }
 
         __attribute__((always_inline)) inline void begin() {
@@ -601,61 +764,93 @@ namespace NeoPixelEx {
             pinMode(kOutputPin, INPUT);
         }
 
+        // clear is equal to
+        // fill(0)
+        // show(0)
         __attribute__((always_inline)) inline void clear() {
             fill(0);
-            _clear(getNumBytes());
+            _clear(_data.size());
+        }
+
+        __attribute__((always_inline)) inline void fill(pixel_type color) {
+            _data.fill(color);
         }
 
         __attribute__((always_inline)) inline void show(uint8_t brightness = 255) {
-            _checkTime();
-            internalShow(kOutputPin, reinterpret_cast<uint8_t *>(data()), getNumBytes(), brightness, &_lastDisplayTime);
+            internalShow(kOutputPin, reinterpret_cast<uint8_t *>(_data.data()), getNumBytes(), brightness, _context);
+        }
+
+        __attribute__((always_inline)) inline pixel_type &operator[](int index) {
+            return data()[index];
+        }
+
+        __attribute__((always_inline)) inline pixel_type operator[](int index) const {
+            return data()[index];
         }
 
         // this method allows to clear any number of pixels
-        __attribute__((always_inline)) void _clear(uint16_t numBytes)
+        // it does not change the actual data
+        __attribute__((always_inline)) void _clear(uint16_t numPixels)
         {
-            _checkTime();
             uint8_t buf[1];
-            internalShow(kOutputPin, buf, numBytes, 0, &_lastDisplayTime);
+            internalShow(kOutputPin, buf, getNumBytes(), 0, _context);
         }
 
-        __attribute__((always_inline)) bool canShow() const {
-            return (micros() > _lastDisplayTime + _Chipset::kMinDisplayPeriod);
+        __attribute__((always_inline)) inline bool canShow() const {
+            return (micros() - _context.getLastDisplayTime() > _Chipset::kMinDisplayPeriod);
         }
 
-    private:
-        __attribute__((always_inline)) void _checkTime() const {
-            uint32_t diff = micros() - _lastDisplayTime;
-            if (diff < _Chipset::kMinDisplayPeriod) {
-                delayMicroseconds(_Chipset::kMinDisplayPeriod - diff);
-            }
+        __attribute__((always_inline)) inline data_type &data() {
+            return _data;
         }
 
-    private:
-        uint32_t _lastDisplayTime;
-    };
+        __attribute__((always_inline)) inline const data_type &data() const {
+            return _data;
+        }
 
     #if NEOPIXEL_HAVE_STATS
-
-        extern Stats stats;
-
-        __attribute__((always_inline)) inline NeoPixelEx::Stats &getStats()
+        __attribute__((always_inline)) inline Stats &getStats()
         {
-            return stats;
+            return _context.getStats();
         }
-
     #endif
+
+    private:
+        data_type _data;
+        Context _context;
+    };
+
+    extern Context _globalContext;
+
+#if NEOPIXEL_HAVE_STATS
+
+    __attribute__((always_inline)) inline Stats &getStats()
+    {
+        return _globalContext.getStats();
+    }
+
+#endif
+
+    inline Context &Context::validate(void *contextPtr)
+    {
+        if (!contextPtr) {
+            return _globalContext;
+        }
+        return *reinterpret_cast<Context *>(contextPtr);
+    }
 
 }
 
 extern "C" {
 
-    inline bool NeoPixel_espShow(uint8_t pin, const uint8_t *pixels, uint16_t numBytes, uint16_t brightness)
+    inline bool NeoPixel_espShow(uint8_t pin, const uint8_t *pixels, uint16_t numBytes, uint16_t brightness, void *contextPtr)
     {
-        uint32_t lastDisplayTime = micros() - 0xffff;
-        return NeoPixelEx::internalShow<NeoPixelEx::DefaultTimings>(pin, pixels, numBytes, brightness, &lastDisplayTime);
+        return NeoPixelEx::internalShow<NeoPixelEx::DefaultTimings>(pin, pixels, numBytes, brightness, NeoPixelEx::Context::validate(contextPtr));
     }
 
 }
 
 #endif
+
+#pragma GCC diagnostic ignored "-Wpragmas" // GCC bug?
+#pragma GCC pop_options